+++ inst/tarpuy/setup.r
@@ -20,14 +20,10 @@ cran <- c(
   , "googleAuthR"
   , "shinydashboard"
   , "stringi"
+  , "bslib"
   )
 
-git <- c(
-  "rstudio/htmltools"
-  , "rstudio/sass"
-  , "rstudio/bslib"
-  , "Flavjack/inti" 
-  )
+git <- c("Flavjack/inti")
 
 for (pkg in cran) { 
   if( !require(pkg, character.only = TRUE) ) {